♻️ Trekk felles funksjoner for V1 og V2 beregning tilsyn barn ut i egen fil
